1. Refactor the import of llava series models. 2. Fix conversation templates issue.
